Ajoute la gestion des états vocaux dans `checkVoiceChat.ts` et améliore le suivi des utilisateurs dans `voiceStateUpdate.ts` avec des logs détaillés.
